main completado

+++ libreria/Funciones.cpp
@@ -2,15 +2,15 @@
 
 float dist(float x1, float y1, float x2, float y2)
 {
-    float x=x2-x1;
-    float y=y2-y1;
-    float distancia=sqrt(x*x+y*y);
+    float x = x2 - x1;
+    float y = y2 - y1;
+    float distancia = sqrt(x * x + y * y);
     return (distancia);
 }
 
-float area(float r)
+double area(float r)
 {
-    float pi=3.141516;
-    float area=pi*(r*r);
+    double pi = 3.141516;
+    double area = pi * (r * r);
     return (area);
 }
